Simplify computations in SVM; simplify plot
Some calculations for generating the kernel
are now performed using `np,dot` instead of
`np.tensordot`.

The inner loop for the statistic analysis
has been removed, since we are no longer
modifying the value of s/gamma. The data
is now visualized as a histogram instead
of a 'continuous boxplot.'

The notebook has been renamed to reflect
its current contents.