Fix various typos in code of Uniform SVM notebook
The deterministic algorithm was training on the test data.
A typo in one of the cells prevented the results from
displaying.

Also, the granularity of the 2D histograms has been
decreased to be able to greater show the patterns
and trends in the data as a whole, rather than the
individual data points.