Expand explanations/analysis in Image_Compression
The explanations in the notebook are fully
revamped to make the notebook accessible to
someone without prior exposure to this particular
topic. The quality of the graphs is vastly
improved as well. They are able to convey much
more information now.

Redundant code was abstracted into a function.

The notebook should be ready for the final
deliverable.