Abstract code; use sklearn RSVD function
The code for testing various low-rank
approximations has been abstracted, so
now it is much easier to add or remove
various approximations without having
to manually add or remove all
references.

Tried using the sklearn RSVD function
instead of our in-house RSVD function
just as a comparison. It has better
accuracy and similar time. Will be
using it from now on for final results,
since it is more reputable as a
public library.

Will probably no longer be using
NPRID and RSVDs, since they are not
the focus of this project.

Various other edits and fixes for
clarity have been made as well.