Make ID testing more robust; fix oversampling
The test to ensure that a given approximation
was formed from an ID has been integrated into
the main computation section, and relevant graphs
are produced as well.

The check_finite parameter in the SciPy QR method
has been set to False to gain improved
performance as suggested in the documentation.

The computations involved in executing OPID with
oversampling have been modified and made more
accurate through the process of adding the robust
ID testing.