Simplify and improve plots of results
Since the last commit, we have been able to
prove that columns are preserved, so we no
longer display a plot of column preservation.
Instead, there is an assert statement in the
testing section that ensures that no strange
errors come up in that regard.

The "entries bounded" variables have been
changed to "max entry" for consistency.

Matplotlib's GridSpec tool is implemented to
more effectively display our results.