Implement many datasets; enhance random sampling
The code to generate many of the datasets that
we want to test on has been added to the
notebook.

As suggested in the NumPy documentation, the
code now uses the Generator class instead of the
RandomState class for random sampling.

The granularity of the set of ranks to test on
has been increased to provide more fine-grained
results.