Fix optim_rid; improve column checking
ropid_ls method has been implemented; it does
not differ in performance much from standard
ropid, so it along with opid have been removed
from the set of methods to be tested.

Since the matrix in optim_rid fails to be positive
definite, the linear system solver has been
switched to one that only assumes that the matrix
is symmetric (which is the case here).

A nested TQDM loop was removed for easier viewing.

A slight code adjustment in optim_rid switched
A[:,cols] to AS[:,_cols]. Mathematically this
should be the same, and performance is unlikely
to be affected. The newer form is more intuitive,
though.

optim_rid occasionally fails column preservation
when tested on the sparse1 dataset, but only by
a very small amount (on the order of 1e-8). A
looser variant of column checking has been added
so that optim_rid can still pass, but whenever
it breaks the stricter limit, a warning message
is outputted.