Scale oversampling with rank; fail on bad data
The default value for the oversampling parameter
is now set to 0.2k instead of 10. This allows
randomized methods to have sufficient entry
bounding on MNIST.

If the rank requested of the ID is greater than
the actual rank of the input data, optim_id now
raises an exception to alert the user that the
input was invalid instead of trying to
compensate and reverting to simpler methods.